+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
   * Show first and last name when adding new time ranges
   * Show kind of time
+  * Support all day events for holidays etc.
+  * Support paging through week view
 
 # Version 1